Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,8 +4,8 @@ float average(int a,int b,int c){
     return avg;
 }
 int main() {
-    int a,b,c;
-    scanf("%d %d %d",&a,&b,&c);
+    float a,b,c;
+    scanf("%f %f %f",&a,&b,&c);
     printf("Average: %.2f",(float)average(a,b,c));
     return 0;
 }
